requirements.txt
asgiref==3.8.1
certifi==2024.2.2
charset-normalizer==3.3.2
Django==5.0.6
easyocr==1.7.1
filelock==3.14.0
fsspec==2024.3.1
idna==3.7
imageio==2.34.1
intel-openmp==2021.4.0
Jinja2==3.1.4
langcodes==3.4.0
langdetect==1.0.9
language_data==1.2.0
lazy_loader==0.4
marisa-trie==1.1.1
MarkupSafe==2.1.5
mkl==2021.4.0
mpmath==1.3.0
networkx==3.3
ninja==1.11.1.1
numpy==1.26.4
opencv-python-headless==4.9.0.80
packaging==24.0
pillow==10.3.0
pyclipper==1.3.0.post5
pytesseract==0.3.10
python-bidi==0.4.2
PyYAML==6.0.1
requests==2.31.0
scikit-image==0.23.2
scipy==1.13.0
shapely==2.0.4
six==1.16.0
sqlparse==0.5.0
sympy==1.12
tbb==2021.12.0
tifffile==2024.5.3
tika==2.6.0
torch==2.3.0
torchvision==0.18.0
typing_extensions==4.11.0
tzdata==2024.1
urllib3==2.2.1
